Move VersionParamType to common module
